login credentials, validation, authorization worked

+++ server/controllers/todo.js
@@ -3,8 +3,11 @@ const { ToDo } = require('../models')
 class ToDoController {
 
   static async findAll(req, res) {
+    const UserId = +req.loginCredential.id
     try {
-      const list = await ToDo.findAll()
+      const list = await ToDo.findAll({
+        where: { UserId }
+      })
       res.status(200).json(list)
     }
     catch (err) {
@@ -13,9 +16,10 @@ class ToDoController {
   }
 
   static async findOne(req, res) {
-    const id = +req.params.id
+    const UserId = +req.loginCredential.id
+    const id = req.params.id
     try {
-      let find = await ToDo.findOne({ where: { id } })
+      let find = await ToDo.findOne({ where: { id, UserId } })
       res.status(200).json(find)
     } catch (err) {
       res.status(500).json(err)
@@ -27,11 +31,14 @@ class ToDoController {
       title: req.body.title,
       description: req.body.description,
       status: req.body.status,
-      due_date: req.body.due_date
+      due_date: req.body.due_date,
+      UserId: +req.loginCredential.id
     }
     console.log(data)
+
     try {
       const add = await ToDo.create(data)
+      console.log(add)
       res.status(201).json(add)
     }
     catch (err) {
@@ -40,11 +47,12 @@ class ToDoController {
   }
 
   static async edit(req, res) {
-    const id = +req.params.id
+    const id = req.params.id
+    const UserId = +req.loginCredential.id
     const { title, description, status, due_date } = req.body
     try {
       const edit = await ToDo.update({ title, description, status, due_date }, {
-        where: { id }, returning: true
+        where: { id, UserId }, returning: true
       })
       res.status(200).json(edit[1][0])
     } catch (err) {
@@ -53,14 +61,17 @@ class ToDoController {
   }
 
   static async editStatus(req, res) {
-    const id = +req.params.id
+    const UserId = +req.loginCredential.id
+    const id = req.params.id
     const status = req.body.status
     try {
-      const editStatus = await ToDo.update(status, {
+      const editStatus = await ToDo.update({ status }, {
         where: {
-          id
+          id,
+          UserId
         }, returning: true
       })
+      console.log(editStatus)
       res.status(200).json(editStatus[1][0])
     } catch (err) {
       res.status(500).json(err)
@@ -68,10 +79,11 @@ class ToDoController {
   }
 
   static async deleted(req, res) {
-    const id = +req.params.id
+    const UserId = +req.loginCredential.id
+    const id = req.params.id
     try {
       const find = await ToDo.findOne({ where: { id } })
-      await ToDo.destroy({ where: { id } }
+      await ToDo.destroy({ where: { id, UserId } }
       )
       res.status(200).json(find)
     } catch (err) {
